\name{internals}
\alias{internals}

\alias{.get_ave_sil_width}
\alias{.get_withinSS}
\alias{print.clustEff}
\alias{print.summary.clustEff}
\alias{print.fpcac}
\alias{print.summary.fpcac}

\title{Internal Functions}
\description{
Functions for internal use only, or not yet documented.
}
\usage{
.get_ave_sil_width(d, cluster)
.get_withinSS(d, cluster)

\method{print}{clustEff}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
\method{print}{summary.clustEff}(x, digits=max(3L, getOption("digits") - 3L), \ldots)

\method{print}{fpcac}(x, digits=max(3L, getOption("digits") - 3L), ...)
\method{print}{summary.fpcac}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
}
\keyword{internal}

