\name{internals}
\alias{internals}
\title{Internal Functions}
\description{
Functions for internal use only, or not yet documented.
}
\usage{
\method{print}{clustEff}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
\method{print}{summary.clustEff}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
}
\keyword{internal}

